Amount is now double

+++ FoodBackend/Food.Server/DishIngredientRelation/DishIngredientResult.cs
@@ -11,7 +11,7 @@ namespace Food.Server.DishIngredientRelation
         public int Id { get; set; }
         public string Name { get; set; }
         public string Description { get; set; }
-        public int Amount { get; set; }
+        public double Amount { get; set; }
         public string Unit { get; set; }
        
     }
